Refactor: minor 오류 해결 위하여 게시판, 댓글 관련 mapper, service 단 수정

+++ happy_diary/src/main/java/com/happydiary/service/CommentServiceImpl.java
@@ -128,7 +128,11 @@ public class CommentServiceImpl implements CommentService {
     public boolean removeCommentOrReply(int cno) {
         boolean successToRemove = true;
         try {
-            commentDao.delete(cno);
+            if (commentDao.select(cno) != null) {
+                commentDao.delete(cno);
+            } else {
+                successToRemove = false;
+            }
         } catch (Exception e) {
             successToRemove = false;
             e.printStackTrace();
